change floor/ceil to floorf/ceilf

+++ mmcv/ops/csrc/onnxruntime/cpu/roi_align.cpp
@@ -144,9 +144,9 @@ void ROIAlignForwardCPU(const int nthreads, const float *input,
     // We use roi_bin_grid to sample the grid and mimic integral
     int roi_bin_grid_h = (sampling_ratio > 0)
                              ? sampling_ratio
-                             : ceil(roi_height / pooled_height);  // e.g., = 2
+                             : ceilf(roi_height / pooled_height);  // e.g., = 2
     int roi_bin_grid_w =
-        (sampling_ratio > 0) ? sampling_ratio : ceil(roi_width / pooled_width);
+        (sampling_ratio > 0) ? sampling_ratio : ceilf(roi_width / pooled_width);
 
     // When the grid is empty, output zeros == 0/1, instead of NaN.
     const float count =
