+++ mmcv/ops/csrc/onnxruntime/cpu/roi_align_rotated.cpp
@@ -152,9 +152,9 @@ void ROIAlignRotatedForwardCPU(const int nthreads, const float *input,
     // We use roi_bin_grid to sample the grid and mimic integral
     int roi_bin_grid_h = (sampling_ratio > 0)
                              ? sampling_ratio
-                             : ceil(roi_height / pooled_height);  // e.g., = 2
+                             : ceilf(roi_height / pooled_height);  // e.g., = 2
     int roi_bin_grid_w =
-        (sampling_ratio > 0) ? sampling_ratio : ceil(roi_width / pooled_width);
+        (sampling_ratio > 0) ? sampling_ratio : ceilf(roi_width / pooled_width);
 
     // We do average (integral) pooling inside a bin
     const float count =
